+++ aaaSystemsCommon/Models/Room.cs
@@ -8,7 +8,7 @@ namespace aaaSystemsCommon.Models
     {
         public int Id { get; set; }
         public long ChatId { get; set; }
-        public long ClientId { get; set; }
+        public int ClientId { get; set; }
         public List<RoomMessage>? RoomMessages { get; set; }
         public List<Participant>? Participants { get; set; }
     }
